Working image to gcode
Convert an image to gcode. Need to clean up and bug test, but the general function seems to work.

For large images, it will likely be better to find the outer contours, and then apply a "fill" for the interior locations. This will fill images but can only draw lines in the 8 cardinal directions.